+++ libs/components/src/lib/cards-container/index.tsx
@@ -6,7 +6,7 @@ export type CardVariantType = keyof CardVariants;
 export type CardVariantProps<T extends CardVariantType> =
   React.ComponentPropsWithoutRef<CardVariants[T]>;
 
-export type CardsContainerCols = 'two' | 'three' | 'four' | 'five';
+export type CardsContainerCols = 'two' | 'three' | 'four' | 'five' | 'infinite';
 
 export interface CardsContainerProps<T extends CardVariantType> {
   variant: T;
@@ -17,10 +17,11 @@ export interface CardsContainerProps<T extends CardVariantType> {
 }
 
 const cardColsVariant: { [key in CardsContainerCols]: string } = {
-  two: 'grid-cols-1 sm:grid-cols-2',
-  three: 'grid-cols-1 sm:grid-cols-2 lg:grid-cols-3',
-  four: 'grid-cols-1 sm:grid-cols-2 lg:grid-cols-4',
-  five: 'grid-cols-1 sm:grid-cols-2 lg:grid-cols-5',
+  two: 'grid grid-cols-1 sm:grid-cols-2',
+  three: 'grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3',
+  four: 'grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4',
+  five: 'grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-5',
+  infinite: 'animate-slide flex',
 };
 
 export const CardsContainer = <T extends CardVariantType>({
@@ -33,20 +34,44 @@ export const CardsContainer = <T extends CardVariantType>({
   const CardComponent = Card[variant];
 
   return (
-    <div
-      className={qtMerge(
-        'grid',
-        'gap-gap-lg', // TODO: we might need to add sm/md/lg/xl variants
-        dense ? 'w-full lg:max-w-max' : 'w-full',
-        cardColsVariant[cols],
-        className,
+    <div className="flex w-full overflow-hidden" id="cards-container">
+      {cols === 'infinite' ? (
+        <>
+          {Array.from(Array(3).keys()).map((index) => (
+            <div
+              key={index}
+              className={qtMerge(
+                'gap-gap-lg', // TODO: Add sm/md/lg/xl variants if needed
+                cardColsVariant[cols],
+                'pr-general-md',
+                className,
+              )}
+              id="infinite-carousel"
+            >
+              {cards.map((card) => (
+                // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+                // @ts-ignore
+                <CardComponent key={card.header} {...card} />
+              ))}
+            </div>
+          ))}
+        </>
+      ) : (
+        <div
+          className={qtMerge(
+            'gap-gap-lg', // TODO: we might need to add sm/md/lg/xl variants
+            dense ? 'w-full lg:max-w-max' : 'w-full',
+            cardColsVariant[cols],
+            className,
+          )}
+        >
+          {cards.map((card) => (
+            // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+            // @ts-ignore
+            <CardComponent key={card.header} {...card} />
+          ))}
+        </div>
       )}
-    >
-      {cards.map((card) => (
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-        // @ts-ignore
-        <CardComponent key={card.header} {...card} />
-      ))}
     </div>
   );
 };