+++ libs/components/CHANGELOG.md
@@ -2,6 +2,37 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.15.0](https://github.com/deriv-com/deriv-com-v2/compare/components-0.14.0...components-0.15.0) (2023-11-30)
+
+## [0.14.0](https://github.com/deriv-com/deriv-com-v2/compare/components-0.13.0...components-0.14.0) (2023-11-30)
+
+## [0.13.0](https://github.com/deriv-com/deriv-com-v2/compare/components-0.12.0...components-0.13.0) (2023-11-30)
+
+
+### Features
+
+* add logo slide carousel in fastpayment ([5cce467](https://github.com/deriv-com/deriv-com-v2/commit/5cce4679d4a6b5259edec915e81e841ff5e06be1))
+
+
+### Bug Fixes
+
+* gap between each array ([ead3532](https://github.com/deriv-com/deriv-com-v2/commit/ead3532498e2e7486dcf9b92dba192b9c5a3866c))
+
+## [0.12.0](https://github.com/deriv-com/deriv-com-v2/compare/components-0.11.0...components-0.12.0) (2023-11-29)
+
+
+### Bug Fixes
+
+* Rating component export issue ([43e0c07](https://github.com/deriv-com/deriv-com-v2/commit/43e0c070ad49d0150397326abc5867cd8433690d))
+
+## [0.11.0](https://github.com/deriv-com/deriv-com-v2/compare/components-0.10.0...components-0.11.0) (2023-11-29)
+
+
+### Bug Fixes
+
+* semantic classname usage ([fa34807](https://github.com/deriv-com/deriv-com-v2/commit/fa34807c97bb18be4fdacfb81b522d4ed52e5fe1))
+* truspilot icons ([672e1dc](https://github.com/deriv-com/deriv-com-v2/commit/672e1dcd8280204d6250ca7da9df1fd8cc5e1f1e))
+
 ## [0.10.0](https://github.com/deriv-com/deriv-com-v2/compare/components-0.9.0...components-0.10.0) (2023-11-28)